test(BridgeGameService): 이동 후 성공 여부를 표시하는 메서드 테스트 추가
